backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
langchain
langchain-community
langchain-google-genai
python-dotenv
openpyxl
pyarrow
fastparquet
python-multipart
